requirements.txt
PyAudio~=0.2.14
requests~=2.32.4
opencv-python~=4.11.0.86
psutil~=7.0.0
dnspython~=2.7.0
PyPDF2~=3.0.1
jwt~=1.4.0
beautifulsoup4~=4.13.4
cryptography~=45.0.5